# 8340490: Shenandoah: Optimize ShenandoahPacer