--- a/src/hotspot/share/gc/shenandoah/shenandoahPacer.cpp
+++ b/src/hotspot/share/gc/shenandoah/shenandoahPacer.cpp
@@ -188,8 +188,8 @@ void ShenandoahPacer::restart_with(size_t non_taxable_bytes, double tax_rate) {
   // Shake up stalled waiters after budget update.
   _need_notify_waiters.try_set();
 }
-
-bool ShenandoahPacer::claim_for_alloc(size_t words, bool force) {
+template<bool FORCE>
+bool ShenandoahPacer::claim_for_alloc(size_t words) {
   assert(ShenandoahPacing, "Only be here when pacing is enabled");
 
   intptr_t tax = MAX2<intptr_t>(1, words * Atomic::load(&_tax_rate));
@@ -198,14 +198,18 @@ bool ShenandoahPacer::claim_for_alloc(size_t words, bool force) {
   intptr_t new_val = 0;
   do {
     cur = Atomic::load(&_budget);
-    if (cur < tax && !force) {
+    if (cur < tax && !FORCE) {
       // Progress depleted, alas.
       return false;
     }
     new_val = cur - tax;
-  } while (Atomic::cmpxchg(&_budget, cur, new_val, memory_order_relaxed) != cur);
+  } while (Atomic::load(&_budget) == cur &&
+           Atomic::cmpxchg(&_budget, cur, new_val, memory_order_relaxed) != cur);
   return true;
 }
+
+template bool ShenandoahPacer::claim_for_alloc<true>(size_t words);
+template bool ShenandoahPacer::claim_for_alloc<false>(size_t words);
 
 void ShenandoahPacer::unpace_for_alloc(intptr_t epoch, size_t words) {
   assert(ShenandoahPacing, "Only be here when pacing is enabled");
@@ -227,17 +231,10 @@ void ShenandoahPacer::pace_for_alloc(size_t words) {
   assert(ShenandoahPacing, "Only be here when pacing is enabled");
 
   // Fast path: try to allocate right away
-  bool claimed = claim_for_alloc(words, false);
+  bool claimed = claim_for_alloc<false>(words);
   if (claimed) {
     return;
   }
-
-  // Forcefully claim the budget: it may go negative at this point, and
-  // GC should replenish for this and subsequent allocations. After this claim,
-  // we would wait a bit until our claim is matched by additional progress,
-  // or the time budget depletes.
-  claimed = claim_for_alloc(words, true);
-  assert(claimed, "Should always succeed");
 
   // Threads that are attaching should not block at all: they are not
   // fully initialized yet. Blocking them would be awkward.
@@ -249,32 +246,40 @@ void ShenandoahPacer::pace_for_alloc(size_t words) {
   JavaThread* current = JavaThread::current();
   if (current->is_attaching_via_jni() ||
       !current->is_active_Java_thread()) {
+    claim_for_alloc<true>(words);
     return;
   }
 
   double start = os::elapsedTime();
+  double end;
 
-  size_t max_ms = ShenandoahPacingMaxDelay;
-  size_t total_ms = 0;
+  size_t const max_ms = ShenandoahPacingMaxDelay;
+  double total_delay = 0;
 
-  while (true) {
+  while (!claimed) {
     // We could instead assist GC, but this would suffice for now.
-    size_t cur_ms = (max_ms > total_ms) ? (max_ms - total_ms) : 1;
-    wait(cur_ms);
+    wait(1);
+    claimed = claim_for_alloc<false>(words);
+    end = os::elapsedTime();
+    total_delay = end - start;
 
-    double end = os::elapsedTime();
-    total_ms = (size_t)((end - start) * 1000);
-
-    if (total_ms > max_ms || Atomic::load(&_budget) >= 0) {
-      // Exiting if either:
-      //  a) Spent local time budget to wait for enough GC progress.
-      //     Breaking out and allocating anyway, which may mean we outpace GC,
-      //     and start Degenerated GC cycle.
-      //  b) The budget had been replenished, which means our claim is satisfied.
-      ShenandoahThreadLocalData::add_paced_time(JavaThread::current(), end - start);
+    if (static_cast<size_t>(total_delay * 1000) > max_ms) {
+      // Exiting if spent local time budget to wait for enough GC progress.
+      // Breaking out and allocating anyway, which may mean we outpace GC,
+      // and start Degenerated GC cycle.
       break;
     }
   }
+
+  if (total_delay > 0) {
+    ShenandoahThreadLocalData::add_paced_time(JavaThread::current(), total_delay);
+  }
+
+  // Forcefully claim if still not claimed after attempts above.
+  if (!claimed) {
+    claimed = claim_for_alloc<true>(words);
+  }
+  assert(claimed, "Should always succeed");
 }
 
 void ShenandoahPacer::wait(size_t time_ms) {