@pengxiaolong 2024-09-20
> I am good with this, assuming performance runs show good results.

Latency wise, in most time it is better than old impl. 

In my specific test with 8G heap on MacOS, throughput is very close to the test w/ ShenandoahPacing disabled, and about 25%~30% improvement comparing the old implementation.