@shipilev 2024-09-21
> > I am good with this, assuming performance runs show good results.
> 
> Latency wise, in most time it is better than old impl.

It is great it improves targeted tests, and it makes sense from the first principles. Run our usual performance pipeline to sanity check if this affects any other benchmarks in any meaningful way.