@pengxiaolong 2024-09-26
> > > I am good with this, assuming performance runs show good results.
> > 
> > 
> > Latency wise, in most time it is better than old impl.
> 
> It is great it improves targeted tests, and it makes sense from the first principles. Run our usual performance pipeline to sanity check if this affects any other benchmarks in any meaningful way.

Performance pipeline showed improvments in most Dacapo benchmarks, we did found very small regression in Dacapo Spring max latency(<1%?), tried to reproduce it with bare metal instance and can't really stably reproduce the regression, sometime  better and sometime worse, it could be just noises.